documentEditorBackend/requirements.txt
# initial write this file

# fastapi
# uvicorn
# sqlalchemy
# passlib[bcrypt]
# python-jose
# requests  # for AI suggestions

# After install above then pip freeze then paste it for future reference of versions
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1